Update rc.lua to make settings not hard-coded
The home_path is needed for the wallpaper code to work, without it someone cloning or forking this project will get the "tried to concatenate a nil value" rc.lua:142 and it'll fallback to the system settings. Also, I used a call to the "config" bath instead of hard coding in your /home/setkah/ or in my case /home/adam folder.